Hotfix: SCN_TO_TIMESTAMP invalid value
We were just too lazy and did not inserted anything to the database in past 5 days or something like that.
Management mode, and the retention times of all flashback archives in the database, but no less than 120 hours. The time for the association to become obsolete elapses only when the database is open. An error is returned if the SCN specified for the argument to SCN_TO_TIMESTAMP is too old
-- https://stackoverflow.com/questions/22681705/how-to-use-timestamp-to-scn-and-scn-to-timestamp-in-oracle